chore : Simply reorganized signup to usercontroller. Not Tested.
