+++ src/main/java/com/de_alone/dokkang/controllers/AuthController.java
@@ -1,16 +1,11 @@
 package com.de_alone.dokkang.controllers;
 
-import com.de_alone.dokkang.models.User;
 import com.de_alone.dokkang.payload.request.LoginRequest;
-import com.de_alone.dokkang.payload.request.SignupRequest;
-import com.de_alone.dokkang.payload.response.MessageResponse;
 import com.de_alone.dokkang.payload.response.LoginResponse;
 import com.de_alone.dokkang.repository.UserRepository;
 import com.de_alone.dokkang.security.jwt.JwtUtils;
 import com.de_alone.dokkang.security.services.UserDetailsImpl;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.HttpHeaders;
-import org.springframework.http.ResponseCookie;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.authentication.AuthenticationManager;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
@@ -52,25 +47,6 @@ public class AuthController {
     return ResponseEntity.ok().body(new LoginResponse(jwtToken));
   }
 
-  @PostMapping("/signup")
-  public ResponseEntity<?> registerUser(@Valid @RequestBody SignupRequest signUpRequest) {
-    if (userRepository.existsByUsername(signUpRequest.getUsername())) {
-      return ResponseEntity.badRequest().body(new MessageResponse("Error: Username is already taken!"));
-    }
 
-    if (userRepository.existsByEmail(signUpRequest.getEmail())) {
-      return ResponseEntity.badRequest().body(new MessageResponse("Error: Email is already in use!"));
-    }
-
-    // Create new user's account
-    User user = new User(signUpRequest.getUsername(),
-                         signUpRequest.getEmail(),
-                         encoder.encode(signUpRequest.getPassword()));
-
-
-    userRepository.save(user);
-
-    return ResponseEntity.ok(new MessageResponse("User registered successfully!"));
-  }
 
 }
