+++ src/main/java/com/de_alone/dokkang/controllers/UserController.java
@@ -1,0 +1,53 @@
+package com.de_alone.dokkang.controllers;
+
+import com.de_alone.dokkang.models.User;
+import com.de_alone.dokkang.payload.request.SignupRequest;
+import com.de_alone.dokkang.payload.response.MessageResponse;
+import com.de_alone.dokkang.repository.UserRepository;
+import com.de_alone.dokkang.security.jwt.JwtUtils;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.ResponseEntity;
+import org.springframework.security.authentication.AuthenticationManager;
+import org.springframework.security.crypto.password.PasswordEncoder;
+import org.springframework.web.bind.annotation.*;
+
+import javax.validation.Valid;
+
+@CrossOrigin(origins = "*", maxAge = 3600)
+@RestController
+@RequestMapping("/auth")
+public class UserController {
+    @Autowired
+    AuthenticationManager authenticationManager;
+
+    @Autowired
+    UserRepository userRepository;
+
+    @Autowired
+    PasswordEncoder encoder;
+
+    @Autowired
+    JwtUtils jwtUtils;
+
+    @PostMapping("/signup")
+    public ResponseEntity<?> registerUser(@Valid @RequestBody SignupRequest signUpRequest) {
+        if (userRepository.existsByUsername(signUpRequest.getUsername())) {
+            return ResponseEntity.badRequest().body(new MessageResponse("Error: Username is already taken!"));
+        }
+
+        if (userRepository.existsByEmail(signUpRequest.getEmail())) {
+            return ResponseEntity.badRequest().body(new MessageResponse("Error: Email is already in use!"));
+        }
+
+        // Create new user's account
+        User user = new User(signUpRequest.getUsername(),
+                signUpRequest.getEmail(),
+                encoder.encode(signUpRequest.getPassword()));
+
+
+        userRepository.save(user);
+
+        //    FIXME 명세에 맞게 수정
+        return ResponseEntity.ok(new MessageResponse("User registered successfully!"));
+    }
+}
